% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import GLKN water data package}
\usage{
importData(type = c("csv"), filepath = NA, protocol = "rivers", new_env = TRUE)
}
\arguments{
\item{type}{Select how to import the data package.
\describe{
\item{"csv"}{Imports the csv version of the data package views from your local machine. If selected, must provide the
filepath for the csvs in the filepath argument.}
\item{"zip"}{Imports the csv version of the data package views from your local machine as a zipped file. If selected, must provide the
filepath and the name of the zip file.}
}}

\item{filepath}{Quoted filepath where data package files live if type = "csv" or type = "zip".}

\item{protocol}{Quoted string of the protocol being imported. For the big rivers protocol, specify "rivers". For the
inland lakes protocol, specify "lakes".}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in GLKN_rivers or GLKN_lakes environment. If \code{FALSE}, stores views in global environment, with
the protocol added as a suffix to the view name.}
}
\value{
Assigns water csvs to specified environment
}
\description{
This function imports views in GLKN water data packages. The easiest way to work with the data package is to
download the published data package from NPS DataStore manually (irma.nps.gov/datastore), or use csvs that match the format
of published data packages. The latter option being for importing pre-certified for tasks like performing QC checks. The other option
is to download a published data package using the NPSutils package. However, the NPSutils package has a lot of dependencies,
including the terra package, which typically fails to install from CRAN. Once NPSutils is installed, the process is less cumbersome.

Each csv (view) in the data package is either added to a GLKN_rivers or GLKN_lakes environment in your workspace, or to your
global environment based on  whether new_env = TRUE or FALSE. If you want to import both rivers and lakes data, you need to run the
importData() function for each, and set new_env = TRUE.
}
\examples{
\dontrun{

# Download lakes and rivers data package from DataStore
devtools::install_github("nationalparkservice/NPSutils") # might take a few tries to install
NPSutils::get_data_packages(c("2306516", "2309369"))

path_lake = "../data/GLKN_water/2306516"
importData(filepath = path_lake, protocol = "lakes")

path_river = "../data/GLKN_water/2309369"
importData(filepath = path_river, protocol = "rivers")

river_zip = ("../data/GLKN_water/records-2306516.zip")
importData(type = 'zip', filepath = river_zip, protocol = 'rivers')



}

}
