% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterBands.R
\name{plotWaterBands}
\alias{plotWaterBands}
\title{plotWaterBands: compare current vs. historic values}
\usage{
plotWaterBands(
  park = "all",
  site = "all",
  site_type = "all",
  year_current = format(Sys.Date(), "\%Y"),
  years_historic = seq(2007, as.numeric(format(Sys.Date(), "\%Y")) - 1, 1),
  active = TRUE,
  parameter = NA,
  include_censored = FALSE,
  sample_depth = "surface",
  legend_position = "none",
  gridlines = "none",
  facet_scales = "fixed"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APIS"}{Apostle Islands National Lakeshore}
\item{"INDU"}{Indiana Dunes National Park}
\item{"ISRO"}{Isle Royale National Park}
\item{"MISS"}{Mississippi National River and Recreation Area}
\item{"PIRO"}{Pictured Rocks National Lakeshore}
\item{"SLBE"}{Sleeping Bear National Lakeshore}
\item{"SACN"}{St. Croix National Scenic Riverway}
\item{"VOYA"}{Voyageurs National Park}
}}

\item{site}{Filter on Location_ID. Easiest way to pick a site. Defaults to "all". Accepted sites are below.
If new sites are added, need to be added to this function as an accepted site.

Rivers: c('INDU_04', 'MISS_UM814', 'MISS_UM822', 'MISS_UM852', 'MISS_UM862', 'MISS_UM868', 'MISS_UM880', 'SACN_APLE_0.5',
          'SACN_KINI_2.2', 'SACN_NAKA_4.8', 'SACN_NAKA_41.3', 'SACN_NAKA_74.5', 'SACN_NAKA_84.6',
          'SACN_SNKE_0.5', 'SACN_STCR_43.7', 'SACN_STCR_63.8', 'SACN_STCR_89.7' 'SACN_STCR_104.0', 'SACN_STCR_138.9')

Lakes: c('APIS_01', 'APIS_02', 'APIS_03', 'APIS_04', 'INDU_01', 'INDU_02', 'INDU_05',
          'ISRO_01', 'ISRO_02', 'ISRO_03', 'ISRO_04', 'ISRO_05', 'ISRO_06', 'ISRO_07', 'ISRO_08', 'ISRO_09',
          'ISRO_13', 'ISRO_14', 'ISRO_19', 'ISRO_20', 'ISRO_21', 'ISRO_22', 'ISRO_24', 'ISRO_30', 'ISRO_32',
          'PIRO_01', 'PIRO_02', 'PIRO_03', 'PIRO_04', 'PIRO_05', 'PIRO_06', 'PIRO_07', 'PIRO_08',
          'SACN_PACQ_SP_01', 'SACN_PHIP_SP_01', 'SACN_STCR_2.0', 'SACN_STCR_15.8', 'SACN_STCR_20.0',
          'SLBE_01', 'SLBE_02', 'SLBE_03', 'SLBE_04', 'SLBE_05', 'SLBE_07', 'SLBE_08', 'SLBE_09', 'SLBE_10', 'SLBE_11',
          'VOYA_01',
          'VOYA_02', 'VOYA_03', 'VOYA_04', 'VOYA_05', 'VOYA_06', 'VOYA_07', 'VOYA_08', 'VOYA_09', 'VOYA_10',
          'VOYA_11', 'VOYA_12', 'VOYA_13', 'VOYA_14', 'VOYA_15', 'VOYA_16', 'VOYA_17', 'VOYA_18', 'VOYA_19',
          'VOYA_20', 'VOYA_21', 'VOYA_22', 'VOYA_23', 'VOYA_24', 'VOYA_25')

Impoundments: c('SACN_CLAM_0.7', 'SACN_STCR_53.9', 'SACN_WILO_0.5')}

\item{site_type}{Quoted string to select either inland lake or stream sites.  Note that impounded sites are not actively
monitored sites, so active = F would need to be specified to return data for those sites. Options are:
\describe{
 \item{"all"}{Includes all location types}
 \item{"lake"}{Location_types that = "Lake"}
 \item{"river"}{Location_types that = "River/Stream"}
 \item{"impound"}{Location_types that = "Riverine Impoundment"}
 }}

\item{year_current}{Year that will be plotted separately. Must be numeric and 4 digits.}

\item{years_historic}{Years to include in historic range calculations. Default starts at 2007.}

\item{active}{Logical. If TRUE (Default) only queries actively monitored sites. If FALSE, returns all sites that have been monitored.}

\item{include_censored}{Logical. If TRUE, the value column includes non-censored and censored values
using the Lower/Upper Quantification Limits values. Censored values are indicated by censored = TRUE, and
are defined as records where the Result_Detection_Condition = Present Above/Below Quantification Limit.
If FALSE (Default), only values with Result_Detection_Condition = "Detected and Quantified"
are returned in the value column. Censored values will be plotted as a star compared with filled circles
for non-censored values.}

\item{sample_depth}{Filter on sample depth. If "all", returns all sample depths. If "surface" (Default),
only returns records that are either blank or "Surface" in the Activity_Relative_Depth column of the Results view.
In general, sample_depth = "surface" is the most appropriate choice.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{facet_scales}{Specify whether facet axes should be "fixed" (all the same) or "free_y", "free_x" or "free" (both).}

\item{sample_type}{Select the sample type. Options available are below. Can only choose one option.
\describe{
\item{"all"}{All possible sampling events.}
\item{"VS"}{Default. GLKN Vital Signs monitoring events, which matches all non-QAQC Activity_Types.}
\item{"QC"}{Includes only include Quality Control sampling events listed in Activity_Type.}
}}

\item{param_name}{Specify the param_name to return. Can only work with 1 param_name at a time. Current accepted values are:.
chemistry: c("ANC_ueqL", "AppColor", "AppColor_PCU", "ChlA_ugL", "Cl_ueqL",
"DOC_mgL", "NH3_mgL", "NO2_mgL", "NO2+NO3_mgL", "NO3_ueqL", "pH_Lab", "PO4_ugL", "SO4_ueqL",
"TN_mgL", "TotDissN_mgL", "TotDissP_ugL", "TP_ugL")
sonde: c("Temp_C", "Temp_F", "SpCond_uScm", "DOsat_pct", "DOsatLoc_pct", "DO_mgL", "pH", "pHmV",
"Turbidity_FNU", "ChlA_RFU", "ChlA_ugL", "BP_mmHg").
other: c("SDepth_m", "Discharge_cfs", "PenetrationRatio", "WaterLevel_Feet", "WaterLevel_m").
Note that "all" is not an accepted value, because there are too many to plot.}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water quality
threshold exists for that param_name and site. If FALSE, no threshold line will be plotted. *NOT CURRENTLY ENABLED*}
}
\value{
Returns a ggplot object of specified current vs historic values
}
\description{
This function produces a plot that summarizes the range of historic data compared with current measurements.
Historic measurements are displayed as min-max values ever previously recorded (outermost band), upper and lower 95%
distribution and middle 50% distribution (inner quartiles) of values previously recorded (inner bands). The line
represents the median value. Sample periods (ie month range on the x-axis) are fixed for each park.
}
\examples{

# RUN IMPORT FIRST: import both lakes and rivers data as zip files
library(waterGLKN)
river_zip = ("../data/GLKN_water/records-2309369.zip")
lake_zip = ("../data/GLKN_water/records-2306516.zip")
importData(type = 'zip', filepath = c(river_zip, lake_zip))

# Plot pH in SACN_STCR_20.0 for 2023
plotWaterBands(site = "SACN_STCR_20.0", year_curr = 2023, years_historic = 2007:2022,
  parameter = "pH", legend_position = 'right')

# Plot ChlA in Lake St. Croix sites for 2023, including censored
lkst <- c("SACN_STCR_20.0", "SACN_STCR_15.8", "SACN_STCR_2.0")
plotWaterBands(site = lkst, year_curr = 2023, years_historic = 2007:2022,
  parameter = "ChlA_ugL", legend_position = 'right', include_censored = T)

# Plot DO in SLBE  2023
plotWaterBands(park = "SLBE", year_curr = 2023, years_historic = 2007:2022,
  parameter = "DOsat_pct", legend_position = 'right')

# Plot ChlA in Lake St. Croix sites with legend on the bottom
lksc <- c('SACN_STCR_2.0', 'SACN_STCR_15.8', 'SACN_STCR_20.0')
plotWaterBands(site = lksc, year_curr = 2023, years_historic = 2007:2022,
               parameter = "ChlA_ugL", legend_position = 'bottom', gridlines = 'grid_y',
               include_censored = T)


}
