% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{plotTrend: Plot WQ trends over time}
\usage{
plotTrend(
  park = "all",
  site = "all",
  site_type = "all",
  sample_type = "VS",
  years = 2007:format(Sys.Date(), "\%Y"),
  months = 4:11,
  active = TRUE,
  parameter = NA,
  sample_depth = "surface",
  include_censored = FALSE,
  layers = c("points", "lines"),
  point_size = 1,
  line_size = 1,
  palette = "viridis",
  smooth = TRUE,
  span = 0.3,
  facet_site = TRUE,
  facet_scales = "free_y",
  legend_position = "none",
  numcol = 2,
  gridlines = "none"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APIS"}{Apostle Islands National Lakeshore}
\item{"INDU"}{Indiana Dunes National Park}
\item{"ISRO"}{Isle Royale National Park}
\item{"MISS"}{Mississippi National River and Recreation Area}
\item{"PIRO"}{Pictured Rocks National Lakeshore}
\item{"SLBE"}{Sleeping Bear National Lakeshore}
\item{"SACN"}{St. Croix National Scenic Riverway}
\item{"VOYA"}{Voyageurs National Park}
}}

\item{site}{Filter on Location_ID. Easiest way to pick a site. Defaults to "all". Accepted sites are below.
If new sites are added, need to be added to this function as an accepted site.

Rivers: c('INDU_04', 'MISS_UM814', 'MISS_UM822', 'MISS_UM852', 'MISS_UM862', 'MISS_UM868', 'MISS_UM880', 'SACN_APLE_0.5',
          'SACN_KINI_2.2', 'SACN_NAKA_4.8', 'SACN_NAKA_41.3', 'SACN_NAKA_74.5', 'SACN_NAKA_84.6',
          'SACN_SNKE_0.5', 'SACN_STCR_43.7', 'SACN_STCR_63.8', 'SACN_STCR_89.7' 'SACN_STCR_104.0', 'SACN_STCR_138.9')

Lakes: c('APIS_01', 'APIS_02', 'APIS_03', 'APIS_04', 'INDU_01', 'INDU_02', 'INDU_05',
          'ISRO_01', 'ISRO_02', 'ISRO_03', 'ISRO_04', 'ISRO_05', 'ISRO_06', 'ISRO_07', 'ISRO_08', 'ISRO_09',
          'ISRO_13', 'ISRO_14', 'ISRO_19', 'ISRO_20', 'ISRO_21', 'ISRO_22', 'ISRO_24', 'ISRO_30', 'ISRO_32',
          'PIRO_01', 'PIRO_02', 'PIRO_03', 'PIRO_04', 'PIRO_05', 'PIRO_06', 'PIRO_07', 'PIRO_08',
          'SACN_PACQ_SP_01', 'SACN_PHIP_SP_01', 'SACN_STCR_2.0', 'SACN_STCR_15.8', 'SACN_STCR_20.0',
          'SLBE_01', 'SLBE_02', 'SLBE_03', 'SLBE_04', 'SLBE_05', 'SLBE_07', 'SLBE_08', 'SLBE_09', 'SLBE_10', 'SLBE_11',
          'VOYA_01',
          'VOYA_02', 'VOYA_03', 'VOYA_04', 'VOYA_05', 'VOYA_06', 'VOYA_07', 'VOYA_08', 'VOYA_09', 'VOYA_10',
          'VOYA_11', 'VOYA_12', 'VOYA_13', 'VOYA_14', 'VOYA_15', 'VOYA_16', 'VOYA_17', 'VOYA_18', 'VOYA_19',
          'VOYA_20', 'VOYA_21', 'VOYA_22', 'VOYA_23', 'VOYA_24', 'VOYA_25')

Impoundments: c('SACN_CLAM_0.7', 'SACN_STCR_53.9', 'SACN_WILO_0.5')}

\item{site_type}{Quoted string to select either inland lake or stream sites.  Note that impounded sites are not actively
monitored sites, so active = F would need to be specified to return data for those sites. Options are:
\describe{
 \item{"all"}{Includes all location types}
 \item{"lake"}{Location_types that = "Lake"}
 \item{"river"}{Location_types that = "River/Stream"}
 \item{"impound"}{Location_types that = "Riverine Impoundment"}
 }}

\item{sample_type}{Select the sample type. Options available are below. Can only choose one option.
\describe{
\item{"all"}{All possible sampling events.}
\item{"VS"}{Default. GLKN Vital Signs monitoring events, which matches all non-QAQC Activity_Types.}
\item{"QC"}{Includes only include Quality Control sampling events listed in Activity_Type.}
}}

\item{years}{Numeric. Years to query. Accepted values start at 2007.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12. Note that most of the
events are between months 4 and 11, and these are set as the defaults.}

\item{active}{Logical. If TRUE (Default), only returns actively monitored locations. If FALSE, returns all
locations that have been monitored at least once since 2007. Active sites are defined as sites that have at
least one sampling event between 2014 and 2024. See "./scripts/active_sites.R" for more details.}

\item{parameter}{Specify one or more parameters to plot. Current accepted values are:
c("Alkalinity_mgL", "Ca_mgL", "ChlA_mgm3", "ChlA_Pheo_pct", "ChlA_ugL", "Cl_mgL", "Depth_m",
  "DO_mgL", "DOC_mgL", "DOsat_pct", "Hg_ngL", "HgMethyl_ngL", "K_mgL", "Mg_mgL", "N_ugL",
  "Na_mgL", "NH4_ugL", "NO2+NO3_ugL", "P_ugL", "pH", "Secchi_m", "Si_mgL", "SO4_mgL", "SpecCond_uScm",
  "TempAir_C", "TempWater_C", "Transp_cm", "TSS_mgL", "Turbidity_NTU", "WaterLevel_m",
  "WaveHt_cm", "WaveHt_m", "WindDir_Deg").}

\item{sample_depth}{Filter on sample depth. If "all" (Default), returns all sample depths. If "surface",
only returns records that are either blank or "Surface" in the Activity_Relative_Depth column of the Results view.
If there are multiple surface samples for a given site, date, parameter, the median value will be used.}

\item{include_censored}{Logical. If TRUE, the value column includes non-censored and censored values
using the Lower/Upper Quantification Limits values. Censored values are indicated by censored = TRUE, and
are defined as records where the Result_Detection_Condition = Present Above/Below Quantification Limit.
If FALSE (Default), only values with Result_Detection_Condition = "Detected and Quantified"
are returned in the value column. Censored values will be plotted as a star compared with filled circles
for non-censored values.}

\item{layers}{Options are "points" and "lines". By default, both will plot.}

\item{point_size}{If specified, will change the size of the points from ggplot default}

\item{line_size}{If specified, will change the size of the line from ggplot default}

\item{palette}{Theme to plot points and lines. Options include 'viridis' (Default- ranges of blue,
green and yellow), magma (yellow, red, purple), plasma (brighter version of magma), turbo (rainbow),
or specify a vector of colors manually. If fewer colors than parameters are specified, they will be
color-ramped to generate enough colors.}

\item{smooth}{Logical. If TRUE (Default), will plot a loess smoothed line. If FALSE, will plot actual line. Only
plots if layers argument includes 'lines'.}

\item{span}{Numeric. Determines how smoothed the line will be for smooth = TRUE. Default is 0.3. Higher spans (up to 1)
cause more smoothing, with 1 being a straight line. Span can range from 0 to 1.}

\item{facet_site}{Logical. If TRUE (default), will facet on site if multiple sites specified. If FALSE, will plot all sites
on the same figure. Only enabled when multiple sites specified.}

\item{facet_scales}{Specify whether facet axes should be fixed (all the same; default) or "free_y", "free_x" or "free" (both).}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple sites are
specified or multiple parks. Default is 2.}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{threshold}{Logical. If TRUE (Default), will plot a dashed (upper) or dotted (lower) line if a water
quality threshold exists for that parameter and site. If FALSE, no threshold line will be plotted.
*NOT CURRENTLY ENABLED*}
}
\value{
Returns a ggplot object of specified parameter trends.
}
\description{
This function produces a line or smoothed trend plot filtered on park, site, year, month, and parameter.
If multiple sites are specified, they will be plotted on the same figure. If multiple parameters are specified,
they will be plotted on separate figures. Note that if you specify a site and parameter combination that doesn't
exist (e.g., a stream site and a parameter only collected in lakes), the function will return an error message
instead of an empty plot.
}
\examples{
\dontrun{

# RUN IMPORT FIRST: import both lakes and rivers data as zip files
river_zip = ("../data/GLKN_water/records-2309369.zip")
lake_zip = ("../data/GLKN_water/records-2306516.zip")
importData(type = 'zip', filepath = c(river_zip, lake_zip))

# Plot smoothed surface pH for ISRO_01 for all years
plotTrend(site = "ISRO_01", parameter = "pH")

# Plot smoothed surface pH for ISRO_01 for all years, removing the legend and using span of 0.75.
plotTrend(site = "ISRO_01", parameter = "pH", span = 0.75)

# Plot smoothed surface ChlA for VOYA_01 for all years including censored values, with palette and larger points/lines.
plotTrend(site = "VOYA_01", parameter = "ChlA_ugL", include_censored = TRUE,
  palette = 'navyblue', point_size = 3, line_size = 1.5)

# Plot non-smoothed DO in SLBE for all years, including the legend,
# different color palette, all sites on the same figure.
plotTrend(park = "SLBE", parameter = "DO_mgL", smooth = F, layers = c('points', 'lines'),
  palette = c("blue", "green", "gold"), point_size = 3, legend_position = 'bottom', facet_site = F)

# Plot smoothed surface SO4 for all VOYA sites all years, including censored, with 0.6 span
plotTrend(park = "VOYA", parameter = "SO4_mgL", include_censored = T, legend_position = "right",
  span = 0.6, point_size = 2.5)

# Plot smoothed surface of multiple Sonde parameters for all PIRO sites over all
# years with 0.6 span.
params <- c("TempWater_C", "SpecCond_uScm", "DOsat_pct", "pH")
plotTrend(park = "PIRO",  facet_site = F,
          parameter = params, legend_position = "right", span = 0.6, palette = 'turbo')

# Plot all \%DO values from ISRO_04, including samples below the surface
plotTrend(site = "ISRO_04", sample_depth = "all", parameter = "DOsat_pct")

}

}
